nix/greeter: update customTheme verification

+++ distro/nix/greeter.nix
@@ -196,7 +196,7 @@ in
 
       if [ -f settings.json ]; then
           theme_file="$(${jq} -r '.customThemeFile // empty' settings.json)"
-          if [ -r "$theme_file" ]; then
+          if [ -f "$theme_file" ] && [ -r "$theme_file" ]; then
               cp "$theme_file" custom-theme.json
               mv settings.json settings.orig.json
               ${jq} '.customThemeFile = "${cacheDir}/custom-theme.json"' settings.orig.json > settings.json
